+++ NEWS.md
@@ -1,3 +1,9 @@
+# utile.tools 0.2.5
+* Added `paste()` & `paste0()`. Base functions with added option for NA removal. Default behavior identical to base functions.
+* Removed `tabulate_model()`. Feature parity has been provided in `utile.tables::build_table()`.
+* Pruned dependency list.
+* Overhauled documentation.
+
 # utile.tools 0.2.4
 * `calc_duration()`: Now supports the return of a `lubridate::duration()` object when the units parameter is left unspecified.